[refactor] #39 alert delegate 패턴으로 변경
